Add static images
