+++ resources/application.conf
@@ -4,6 +4,6 @@ ktor {
         host = 127.0.0.1
     }
     application {
-        modules = [ io.krugosvet.dailydish.ApplicationKt.module ]
+        modules = [ io.krugosvet.dailydish.ApplicationKt.main, io.krugosvet.dailydish.MealModuleKt.module ]
     }
 }
